+++ scheduler/printers.c
@@ -645,6 +645,12 @@ cupsdDeletePrinter(
 		              "Job stopped.");
 
  /*
+  * Expire subscriptions on the printer...
+  */
+
+  cupsdExpireSubscriptions(p, /*job*/NULL);
+
+ /*
   * Remove the printer from the list...
   */